wikifood/requirements.txt
cachelib==0.4.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
distlib==0.3.3
filelock==3.1.0
Flask==2.0.2
gcloud==0.18.3
googleapis-common-protos==1.54.0
greenlet==1.1.1
gunicorn==20.1.0
httplib2==0.20.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jwcrypto==1.0
MarkupSafe==2.0.1
multidict==5.1.0
mysql-connector==2.2.9
numpy==1.22.3
oauth2client==4.1.3
pbr==5.8.1
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
pyparsing==3.0.6
Pyrebase4==4.5.0
python-jwt==3.3.2
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
termcolor==1.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uuid==1.30
webptools==0.0.7
Werkzeug==2.0.1
wrapt==1.13.3
yarl==1.6.3